Move hooks to their own file
