+++ .eslintrc
@@ -1,9 +1,3 @@
 {
-  "extends": "vtex",
-  "root": true,
-  "env": {
-    "node": true,
-    "es6": true,
-    "jest": true
-  }
+  "extends": "vtex"
 }
