+++ package.json
@@ -21,15 +21,15 @@
     ]
   },
   "devDependencies": {
-    "@types/node": "^12.7.12",
-    "@vtex/prettier-config": "^0.1.4",
-    "eslint": "^6.8.0",
-    "eslint-config-vtex": "^12.3.2",
-    "eslint-config-vtex-react": "^6.3.2",
-    "husky": "^4.2.0",
-    "lint-staged": "^10.0.2",
-    "prettier": "^2.0.4",
-    "react": "^16.12.0",
-    "typescript": "^3.8.3"
+    "@types/node": "^14.6.2",
+    "@vtex/prettier-config": "^0.3.1",
+    "eslint": "^7.8.0",
+    "eslint-config-vtex": "^12.8.6",
+    "eslint-config-vtex-react": "^6.7.8",
+    "husky": "^4.2.5",
+    "lint-staged": "^10.2.13",
+    "prettier": "^2.1.1",
+    "react": "^16.13.1",
+    "typescript": "^4.0.2"
   }
 }
