+++ react/.eslintrc
@@ -1,7 +1,3 @@
 {
-  "extends": "vtex-react",
-  "env": {
-    "browser": true,
-    "es6": true
-  }
+  "extends": "vtex-react"
 }
